LPD-71919 Improve Performance
